requirements.txt
Django
djangorestframework
elasticsearch
PyJWT
python-dotenv
djangorestframework-simplejwt
django-cors-headers